Update skeleton of All files in the AI src folder
Finished the .cpp files in the AI folder. This is where I'll move all movement
of objects. Eventually, I'd want to create some form of simple anim tree which
will change the state of each object. The AI manager will determine if an
object's state is changed and will attach new algorithms to the AI Component
before it runs it. Optimizations would probably include creating a update list
or vector where whenever the state of an object is created, it's AI component
is appended to the update list and the AI manager will just go through the list
and change stuff when necessary.